refactor: header 컴포넌트 분리 및 컴포넌트 폴더 정리
