+++ src/app/order/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import Header2 from "@/components/Header2";
+import HeaderToBack from "@/components/ui/Headers/HeaderToBack";
 import DeliveryAddress from "@/components/order/DeliveryAddress";
 import DeliveryRequest from "@/components/order/DeliveryRequest";
 import TypeOfPayment from "@/components/order/TypeOfPayment";
@@ -8,14 +8,14 @@ import ExpectedPaymoney from "@/components/order/ExpectedPaymoney";
 import TermsOfService from "@/components/order/TermsOfService";
 import InformationOfOrderer from "@/components/order/InformationOfOrderer";
 import DeliveryItemList from "@/components/order/DeliveryItemList";
-import ButtonOfOrder from "@/components/ui/ButtonOfOrder";
+import ButtonOfOrder from "@/components/ui/Buttons/ButtonOfOrder";
 
 export default function OrderPage(){
 
 
     return(
         <>
-            <Header2 />
+            <HeaderToBack title="결제하기" />
             <div className="bg-[#f5f5f5]">
                 <DeliveryAddress />
                 <DeliveryRequest />
