+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import BottomNav from "@/components/BottomNav";
 import QuickMenu from "@/components/QuickMenu";
-import SearchBar from "@/components/SearchBar";
+import SearchBar from "@/components/ui/Headers/HeaderWithSearchBar";
 import TopNav from "@/components/TopNav";
 import AdBox from "@/components/ui/AdBox";
 import Swiper from "@/components/ui/Swiper";
