+++ src/components/cart/CartList.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react'
 import Image from 'next/image'
 import trash from '@/asset/images/Trash.png'
 import '@/app/cart/cart.css'
-import Buttons from '@/components/ui/buttons'
+import Buttons from '@/components/ui/Buttons/buttons'
 // import { useRecoilState } from 'recoil'
 // import { CartCheckedListAtom } from '@/state/CartCheckedListAtom'
 import { productData } from '@/lib/CartList'
