+++ src/components/ui/Headers/HeaderWithSearchBar.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 
-export default function SearchBar() {
+export default function HeaderWithSearchBar() {
 
     return (
         <>
